Rename Translation.ts to useTranslation.ts
