+++ ui/appui-react/src/appui-react/accudraw/AccuDrawDialog.tsx
@@ -14,7 +14,7 @@ import { Orientation } from "@itwin/core-react";
 import { UiFramework } from "../UiFramework";
 import { ModelessDialog } from "../dialog/ModelessDialog";
 import { AccuDrawFieldContainer } from "./AccuDrawFieldContainer";
-import { useTranslation } from "../Translation";
+import { useTranslation } from "../useTranslation";
 
 /** Properties for [[AccuDrawDialog]]
  * @beta */
