+++ ui/appui-react/src/appui-react/messages/MessageRenderer.tsx
@@ -13,7 +13,7 @@ import { ProgressLinear, Text, useToaster } from "@itwin/itwinui-react";
 import type { ActivityMessageEventArgs } from "../messages/MessageManager";
 import { MessageManager } from "../messages/MessageManager";
 import { OutputMessagePriority, OutputMessageType } from "@itwin/core-frontend";
-import { useTranslation } from "../Translation";
+import { useTranslation } from "../useTranslation";
 
 interface UseActivityMessageProps {
   cancelActivityMessage?: () => void;
