+++ ui/appui-react/src/appui-react/settings/quantityformatting/UnitSystemSelector.tsx
@@ -10,7 +10,7 @@ import * as React from "react";
 import type { UnitSystemKey } from "@itwin/core-quantity";
 import type { SelectOption } from "@itwin/itwinui-react";
 import { Select } from "@itwin/itwinui-react";
-import { useTranslation } from "../../Translation";
+import { useTranslation } from "../../useTranslation";
 
 /** Props for [[UnitSystemSelector]]
  * @beta
