+++ ui/appui-react/src/appui-react/settings/ui/UiSettingsPage.tsx
@@ -24,7 +24,7 @@ import {
 import type { SelectOption } from "@itwin/itwinui-react";
 import { Select, Slider, ToggleSwitch } from "@itwin/itwinui-react";
 import type { UiSyncEventArgs } from "../../syncui/UiSyncEvent";
-import { useTranslation } from "../../Translation";
+import { useTranslation } from "../../useTranslation";
 
 /** UiSettingsPage displaying the active UI settings. This page lets users set the following settings.
  *
