+++ ui/appui-react/src/appui-react/statusbar/Overflow.tsx
@@ -12,7 +12,7 @@ import classnames from "classnames";
 import * as React from "react";
 import { Ellipsis } from "../layout/base/Ellipsis";
 import "./Overflow.scss";
-import { useTranslation } from "../Translation";
+import { useTranslation } from "../useTranslation";
 
 /** Properties of [[StatusBarOverflow]] component.
  * @internal
