+++ ui/components-react/src/components-react/filter-builder/FilterBuilderToolbar.tsx
@@ -8,7 +8,7 @@
 import React from "react";
 import { Flex, IconButton } from "@itwin/itwinui-react";
 import { SvgAdd, SvgDelete } from "@itwin/itwinui-icons-react";
-import { useTranslation } from "../Translation";
+import { useTranslation } from "../useTranslation";
 /**
  * Props for [[PropertyFilterBuilderToolbar]] component.
  * @internal
