+++ ui/components-react/src/components-react/toolbar/InternalToolbarComponent.tsx
@@ -40,7 +40,7 @@ import {
 import { SvgPlaceholder } from "@itwin/itwinui-icons-react";
 import type { BeEvent } from "@itwin/core-bentley";
 import { useConditionalSynchedItems } from "./useConditionalSynchedItems";
-import { useTranslation } from "../Translation";
+import { useTranslation } from "../useTranslation";
 
 /** Describes the data needed to insert a custom `React` button into an ToolbarWithOverflow.
  * @public
